% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_results.R
\name{access_results}
\alias{access_results}
\title{Access FP results}
\usage{
access_results(fp_results, regions)
}
\arguments{
\item{fp_results}{tibble with iso (to filter on)}

\item{regions}{subset of fp_results$iso (error is returned when region is not included)}
}
\value{
fp_results for the selected regions
}
\description{
Access FP results
}
