% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_regional_results.R
\name{create_regional_results}
\alias{create_regional_results}
\title{Title: Summarize regional samples}
\usage{
create_regional_results(samples)
}
\arguments{
\item{samples}{tibble or data frame with iso-year-samples-region-population counts
(iso refers to country iso, variable names for samples include "samp")}
}
\value{
tibble with region-year-median-lower-upper
}
\description{
Summarize regional samples into points estimates (median), and
uncertainty intervals (lower and upper given by the 2.5th and 97.5th percentiles
of the sample, respectively)
}
\examples{
create_regional_results(tribble(
  ~iso, ~ year, ~ sample1, ~ sample2, ~region, ~population_ct,
     4, 1990, 0.3, 0.4, "Southern Asia", 2025079,
     4, 1991, 0.5, 0.6, "Southern Asia", 2069776,
     858, 2020, 0.7, 0.9, "South America", 440817))

}
