% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_fp.R
\name{summarize_fp}
\alias{summarize_fp}
\title{Summarize samples per area-year}
\usage{
summarize_fp(samples)
}
\arguments{
\item{samples}{tibble or data frame with iso-year-samples (iso refers to country iso or region name,
variable names for samples include "samp")}
}
\value{
tibble with iso-year-median-lower-upper
}
\description{
Summarize samples per area-year into median, lower and upper, where lower and
upper are given by the 2.5th and 97.5th percentiles of the sample,
respectively
}
\examples{
 summarize_fp(tribble(
 ~iso, ~ year, ~ sample1, ~ sample2,
 4, 1990, 0.5, 0.7,
 4, 1990, 0.4, 0.8,
 5, 1991, 0.3, 0.2))
}
